api/requirements.txt
Flask
requests
jinja2
pyjson
pyrandom
rdflib
pyld
qrcode
Pillow